requirements.txt
streamlit
langchain
langchain-openai
langgraph
python-dotenv
openai